visibility folder migrated to ts
