+++ tests/admin/services/visibility/service-overview.visibility.test.ts
@@ -1,0 +1,31 @@
+import { ServicesOverviewPage } from '@page-objects/services/overview/services-overview-page';
+
+import { test } from '@setup/test-setup';
+import { URLS } from '@utils/env/urls';
+
+test('[services] [visibility] Service overview page elements visibility', async ({ page }) => {
+  await page.goto(URLS.admin + 'services/overview');
+  await page.waitForLoadState('domcontentloaded');
+
+  const sop = new ServicesOverviewPage(page);
+
+  await test.step('Service name column/values visible', async () => {
+    await sop.assertServiceNameExists();
+  });
+
+  await test.step('Service status column/values visible', async () => {
+    await sop.assertStatusExists();
+  });
+
+  await test.step('Edit button visible', async () => {
+    await sop.assertEditButtonExists();
+  });
+
+  await test.step('Delete button visible', async () => {
+    await sop.assertDeleteButtonExists();
+  });
+
+  await test.step('Pagination / page size control visible', async () => {
+    await sop.assertPageSizeVisibleServices();
+  });
+});
